TSDK-872 Fix IT
