+++ public-api/src/main/scala/xyz/stratalab/bridge/publicapi/PublicApiHttpServiceServer.scala
@@ -126,7 +126,7 @@ object PublicApiHttpServiceServer {
               "Error starting pegin session"
             )
         }
-      case req @ POST -> Root / BridgeContants.TOPL_MINTING_STATUS =>
+      case req @ POST -> Root / BridgeContants.STRATA_MINTING_STATUS =>
         implicit val mintingStatusRequestDecoder
             : EntityDecoder[IO, MintingStatusRequest] =
           jsonOf[IO, MintingStatusRequest]
