+++ shared/src/main/scala/xyz/stratalab/bridge/shared/BridgeContants.scala
@@ -5,6 +5,6 @@ object BridgeContants {
 
   val START_PEGOUT_SESSION_PATH = "start-session-pegout"
 
-  val TOPL_MINTING_STATUS = "strata-minting-status"
+  val STRATA_MINTING_STATUS = "strata-minting-status"
 
 }
